Reorganising code....
run/scripting  to handle the lifecycle.
